App should have what it needs to have
