+++ client/src/App.js
@@ -1,21 +1,19 @@
 import React, { useState, useEffect } from 'react';
+import { Route, Switch, useHistory } from 'react-router-dom';
 import Header from './components/shared/Header/Header';
 import Main from './components/main/Main';
-import Footer from './components/shared/Footer/Footer';
 import Login from './components/auth/login/Login';
 import Register from './components/auth/register/Register'
-import { Route, Switch, useHistory } from 'react-router-dom';
+import MainAuth from './components/main/MainAuth'
+import Footer from './components/shared/Footer/Footer';
 import { loginUser, registerUser, removeToken, verifyUser } from './services/auth';
-import { getAllProducts } from '../../services/products'
-
 import './App.css';
 
 function App() {
   const [currentUser, setCurrentUser] = useState(null);
-  const [allProducts, setAllProducts] = useState([])
   const history = useHistory();
 
-// LOGIN STUFF //
+///LOGIN STUFF///
   useEffect(() => {
     const handleVerify = async () => {
       const userData = await verifyUser();
@@ -41,48 +39,28 @@ function App() {
     localStorage.removeItem('authToken');
     removeToken();
   }
-  //////////////////////////////
-  useEffect(() => {
-    const fetchProducts = async () => {
-      const products = await getAllProducts()
-      setAllProducts(products)
-      setQueriedProducts(products)
-    }
-    fetchProducts()
-  }, [])
 
-  const handleSubmit = event => event.preventDefault()
+
 
   return (
     <div className="App">
-      <Header />
+      <Header
+        currentUser={currentUser}
+        handleLogout={handleLogout}
+      />
       <Switch>
+        <Route path='/'>
+          <Main />
+        </Route>
         <Route path='/login'>
           <Login handleLogin={handleLogin} />
         </Route>
         <Route path='/register'>
           <Register handleRegister={handleRegister} />
         </Route>
-        <Route path='/'>
-          <Main>
-            currentUser={currentUser}
-          handleLogout={handleLogout}
-          </Main>
+        <Route path='/auth/main'>
+          <MainAuth />
         </Route>
-        {/* <Route> */}
-        {/* <CreateSO
-        onSubmit={handleSubmit} /> */}
-        
-        {/* </Route> */}
-        {/* <Route> */}
-          {/* <DetailsSO /> */}
-        {/* </Route> */}
-        {/* <Route> */}
-          {/* <Products /> */}
-        {/* </Route> */}
-        {/* <Route> */}
-          {/* <UpdateSO /> */}
-        {/* </Route> */}
       </Switch>
       <Footer />
     </div >
