+++ client/src/components/main/MainAuth.jsx
@@ -1,22 +1,48 @@
-import React from 'react';
-import { Link } from "react-router-dom";
-import CreateSO from '../screens/createSo/CreateSo';
+import React, { useState, useEffect } from 'react';
+// import { Link } from "react-router-dom";
+// import CreateSO from '../screens/createSo/CreateSo';
+import { getAllProducts } from '../../services/products'
 
 
-const MainAuth = (props) => {
 
-  const salesOrdersMapped = SalesOrders.map((salesOrder, index) =>
-  <SalesOrders _id={salesOrder._id} name={product.name} price={product.price} key={index} /> THIS IS NOT RITE
-)
+const MainAuth = () => {
+  const [allProducts, setAllProducts] = useState([]);
+
+  useEffect(() => {
+    const fetchProducts = async () => {
+      const products = await getAllProducts()
+      setAllProducts(products);
+      console.log(setAllProducts)
+    };
+    fetchProducts();
+  }, []);
+
+//   const salesOrdersMapped = SalesOrders.map((salesOrder, index) =>
+//   <SalesOrders _id={salesOrder._id} name={product.name} price={product.price} key={index} /> 
+// )
   // add state for current user
   return (
     <div class="main-container">
-      <h1 id="main-title">Da Haus.</h1>
+      <h1 id="main-title">This is Main Auth.</h1>
       <h2 id="main-subtitle">Welcome, </h2>
       <div id="main-subcontainer">
-        {salesOrdersMapped}
+        {/* {salesOrdersMapped} */}
         </div>
-      </div>
+    </div>
+    {/* <Route> */}
+    <CreateSO
+    onSubmit={handleSubmit} /> 
+    
+    {/* </Route> */}
+    {/* <Route> */}
+      {/* <DetailsSO /> */}
+    {/* </Route> */}
+    {/* <Route> */}
+      {/* <Products /> */}
+    {/* </Route> */}
+    {/* <Route> */}
+      {/* <UpdateSO /> */}
+    {/* </Route> */}
   );
 };
