+++ client/src/components/screens/createSO/CreateSO.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { Link } from "react-router-dom";
-import Products from '../screens/products/Products';
+// import { Link } from "react-router-dom";
+// import Products from '../screens/products/Products';
 
 const CreateSO = (props) => {
 
-  const productsMapped = Products.map((product, index) =>
-  <Product _id={product._id} name={product.name} price={product.price} key={index} />
-)
+//   const productsMapped = Products.map((product, index) =>
+//   <Product _id={product._id} name={product.name} price={product.price} key={index} />
+// )
   // add state for current user
   return (
     <div class="main-container">
@@ -14,11 +14,11 @@ const CreateSO = (props) => {
       <h2 id="main-subtitle">Welcome, </h2>
       <div id="main-subcontainer">
         <div id="right-side">
-        {
+        {/* {
         Products.map((product, index) => (
         />
           ))
-        }
+        } */}
         </div>
       </div>
     </div>
